+++ apps/foh/CHANGELOG.md
@@ -1,5 +1,12 @@
 # foh
 
+## 0.5.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - lib@0.3.1
+
 ## 0.5.3
 
 ### Patch Changes